refactor: Separate accounts and transactions logic into different apps
