+++ accounts/README.md
@@ -1,0 +1,3 @@
+# Accounts  API Endpoints
+
+This directory contains the API endpoints for the accounts microservice.
