+++ accounts/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AccountsTransactionsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "accounts_transactions"
+    name = "accounts"
