--- accounts_transactions/README.md
@@ -1,4 +1,0 @@
-# Accounts and Transactions API Endpoints
-
-This directory contains the API endpoints for the accounts and transactions
-microservice.
